Added Some Functionality to Login and Signup Page
The Login page does a POST request to the server and does validation on the data before entering it into the database. The Signup page does the same. Right now, there is not way to sign up because the email field doesn't exist, and the Schema needs to have an email to be validated by Mongoose. Currently, I had to move all the routing information into one file - I will find a way to split them up at some point.