User Login and Session Updated.
Changed how connects are made the MongoDB. Instead of terminating the connection to the cluster every time a request is made, the connection pool stays open for any requested database. Users can now log in and see their profile. Routing information updated so that there are no pathological outcomes (aka. user logging in when they are logged in) etc.