Initial Structure for Non-Hierarchical Permissions
Added an accounts.js file to enumerate specific accounts - the necessity was that handling accounts as strings is too messy and prone to error.

Added a permissions.js page - this file handles the set of permissions for each route. Permission control is done by string, because this is the easiest way without setting up chaotic infrastructure. Just have to be careful.
All routes are now set with the permissions.

Still working to fix authorization.js