update articles.accessor.js using the sortArticles method
the sortArticles method follows the structure of previous methods and attempts to sort the articles from oldest to newest. It starts by defining parameters for the function and then sets a try catch block to catch any errors during the execution of the code inside. It then opens up a connection and waits it to be established. The if else block checks the sorting order. If the order is ascending, then the sortOrder is set to 1, else it's set to -1. The constant skip is used for pagination and the constant articles finds all the articles, sorts them by the sortBy field in sortOrder, skips certain articles, and then limits the number of articles on a certain page to improve speed. Finally, the function just returns the sorted articles.