Delete PhotoTags given tag name
Fixed the paths to photo model as there were some errors. Allows users with admin or photographer authorization to delete a photo tag given the name of the photo tag in the url. Deletion of a nonexistent tag would throw an error. Deletion of an existing tag will have cascading effects to models with references to photoTags-- which at this moment, only the photo model has a reference to photoTag. Three tests are provided:  proper deletion, improper deletion, proper deletion with cascading effect.